Crate moxie_dom

Tools for declaratively constructing and incrementally updating HTML DOM trees on the web. Based on the [moxie] UI runtime.

Re-exports

pub use augdom::sys;

Modules

elements
Element definitions generated from the listing on MDN. Because topologically-nested functions are called-by-macro today, the element macros generated here appear in the root module.

embed
Embedding APIs offering finer-grained control over execution of the runtime.

event
Event types.

rsdom
An implementation of augdom's APIs on top of an in-memory emulation of the web's DOM.

sys
Raw API bindings for Web APIs

webdom
An implementation of augdom's APIs on top of the actual web using the web-sys crate and wasm-bindgen.

Macros

a
The HTML <a> element (or anchor element), along with its href attribute, creates a hyperlink to other web pages, files, locations within the same page, email addresses, or any other URL.

abbr
The HTML Abbreviation element (<abbr>) represents an abbreviation or acronym; the optional title attribute can provide an expansion or description for the abbreviation.

address
The HTML <address> element indicates that the enclosed HTML provides contact information for a person or people, or for an organization.

area
The HTML <area> element defines a hot-spot region on an image, and optionally associates it with a hypertext link. This element is used only within a <map> element.

article
The HTML <article> element represents a self-contained composition in a document, page, application, or site, which is intended to be independently distributable or reusable (e.g., in syndication).

aside
The HTML <aside> element represents a portion of a document whose content is only indirectly related to the document's main content.

audio
The HTML <audio> element is used to embed sound content in documents. It may contain one or more audio sources, represented using the src attribute or the <source> element: the browser will choose the most suitable one. It can also be the destination for streamed media, using a MediaStream.

b
The HTML Bring Attention To element (<b>) is used to draw the reader's attention to the element's contents, which are not otherwise granted special importance.

base
The HTML <base> element specifies the base URL to use for all relative URLs contained within a document. There can be only one <base> element in a document.

bdi
The HTML Bidirectional Isolate element (<bdi>) tells the browser's bidirectional algorithm to treat the text it contains in isolation from its surrounding text.

bdo
The HTML Bidirectional Text Override element (<bdo>) overrides the current directionality of text, so that the text within is rendered in a different direction.

blockquote
The HTML <blockquote> element (or HTML Block Quotation Element) indicates that the enclosed text is an extended quotation. Usually, this is rendered visually by indentation. A URL for the source of the quotation may be given using the cite attribute, while a text representation of the source can be given using the <cite> element.

body
The HTML <body> element represents the content of an HTML document. There can be only one <body> element in a document.

br
The HTML <br> element produces a line break in text (carriage-return). It is useful for writing a poem or an address, where the division of lines is significant.

button
The HTML <button> element represents a clickable button, which can be used in forms or anywhere in a document that needs simple, standard button functionality.

canvas
Use the HTML <canvas> element with either the canvas scripting API or the WebGL API to draw graphics and animations.

caption
The HTML Table Caption element (<caption>) specifies the caption (or title) of a table, and if used is always the first child of a <table>.

cite
The HTML Citation element (<cite>) is used to describe a reference to a cited creative work, and must include the title of that work.

code
The HTML <code> element displays its contents styled in a fashion intended to indicate that the text is a short fragment of computer code.

col
The HTML <col> element defines a column within a table and is used for defining common semantics on all common cells. It is generally found within a <colgroup> element.

colgroup
The HTML <colgroup> element defines a group of columns within a table.

data
The HTML <data> element links a given content with a machine-readable translation. If the content is time- or date-related, the <time> element must be used.

datalist
The HTML <datalist> element contains a set of <option> elements that represent the values available for other controls.

dd
The HTML <dd> element provides the description, definition, or value for the preceding term (<dt>) in a description list (<dl>).

del
The HTML <del> element represents a range of text that has been deleted from a document.

details
The HTML Details Element (<details>) creates a disclosure widget in which information is visible only when the widget is toggled into an "open" state.

dfn
The HTML Definition element (<dfn>) is used to indicate the term being defined within the context of a definition phrase or sentence.

dialog
The HTML <dialog> element represents a dialog box or other interactive component, such as an inspector or window.

div
The HTML Content Division element (<div>) is the generic container for flow content. It has no effect on the content or layout until styled using CSS.

dl
The HTML <dl> element represents a description list. The element encloses a list of groups of terms (specified using the <dt> element) and descriptions (provided by <dd> elements). Common uses for this element are to implement a glossary or to display metadata (a list of key-value pairs).

dt
The HTML <dt> element specifies a term in a description or definition list, and as such must be used inside a <dl> element.

element
Create and mount an HTML element. Called from the individual element macros, which in turn are normally called by the output of the [moxie::mox!] macro.

em
The HTML <em> element marks text that has stress emphasis. The <em> element can be nested, with each level of nesting indicating a greater degree of emphasis.

embed
The HTML <embed> element embeds external content at the specified point in the document. This content is provided by an external application or other source of interactive content such as a browser plug-in.

fieldset
The HTML <fieldset> element is used to group several controls as well as labels ([<label>][label]) within a web form.

figcaption
The HTML <figcaption> or Figure Caption element represents a caption or legend describing the rest of the contents of its parent <figure> element.

figure
The HTML <figure> (Figure With Optional Caption) element represents self-contained content, potentially with an optional caption, which is specified using the (<figcaption>) element.

footer
The HTML <footer> element represents a footer for its nearest sectioning content or sectioning root element. A footer typically contains information about the author of the section, copyright data or links to related documents.

form
The HTML <form> element represents a document section that contains interactive controls for submitting information to a web server.

h1
The HTML <h1>–<h6> elements represent six levels of section headings. <h1> is the highest section level and <h6> is the lowest.

h2
The HTML <h1>–<h6> elements represent six levels of section headings. <h1> is the highest section level and <h6> is the lowest.

h3
The HTML <h1>–<h6> elements represent six levels of section headings. <h1> is the highest section level and <h6> is the lowest.

h4
The HTML <h1>–<h6> elements represent six levels of section headings. <h1> is the highest section level and <h6> is the lowest.

h5
The HTML <h1>–<h6> elements represent six levels of section headings. <h1> is the highest section level and <h6> is the lowest.

h6
The HTML <h1>–<h6> elements represent six levels of section headings. <h1> is the highest section level and <h6> is the lowest.

head
The HTML <head> element contains machine-readable information (metadata) about the document, like its title, scripts, and style sheets.

header
The HTML <header> element represents introductory content, typically a group of introductory or navigational aids. It may contain some heading elements but also a logo, a search form, an author name, and other elements.

hgroup
The [HTML <hgroup> element][mdn] represents a multi-level heading for a section of a document. It groups a set of [<h1>–<h6>][heading] elements.

hr
The HTML <hr> element represents a thematic break between paragraph-level elements: for example, a change of scene in a story, or a shift of topic within a section.

html
The <html> element represents the root (top-level element) of an HTML document, so it is also referred to as the root element. All other elements must be descendants of this element.

i
The HTML <i> element represents a range of text that is set off from the normal text for some reason. Some examples include technical terms, foreign language phrases, or fictional character thoughts. It is typically displayed in italic type.

iframe
The HTML Inline Frame element (<iframe>) represents a nested browsing context, embedding another HTML page into the current one.

img
The HTML <img> element embeds an image into the document.

input
The HTML <input> element is used to create interactive controls for web-based forms in order to accept data from the user; a wide variety of types of input data and control widgets are available, depending on the device and user agent.

ins
The HTML <ins> element represents a range of text that has been added to a document.

kbd
The HTML Keyboard Input element (<kbd>) represents a span of inline text denoting textual user input from a keyboard, voice input, or any other text entry device.

label
The HTML <label> element represents a caption for an item in a user interface.

legend
The HTML <legend> element represents a caption for the content of its parent <fieldset>.

li
The HTML <li> element is used to represent an item in a list.

link
The HTML External Resource Link element (<link>) specifies relationships between the current document and an external resource. This element is most commonly used to link to stylesheets, but is also used to establish site icons (both "favicon" style icons and icons for the home screen and apps on mobile devices) among other things.

main
The HTML <main> element represents the dominant content of the <body> of a document. The main content area consists of content that is directly related to or expands upon the central topic of a document, or the central functionality of an application.

map
The HTML <map> element is used with <area> elements to define an image map (a clickable link area).

mark
The HTML Mark Text element (<mark>) represents text which is marked or highlighted for reference or notation purposes, due to the marked passage's relevance or importance in the enclosing context.

menu
The HTML <menu> element represents a group of commands that a user can perform or activate. This includes both list menus, which might appear across the top of a screen, as well as context menus, such as those that might appear underneath a button after it has been clicked.

meta
The HTML <meta> element represents metadata that cannot be represented by other HTML meta-related elements, like [<base>], [<link>], [<script>], [<style>] or [<title>].

meter
The HTML <meter> element represents either a scalar value within a known range or a fractional value.

nav
The HTML <nav> element represents a section of a page whose purpose is to provide navigation links, either within the current document or to other documents. Common examples of navigation sections are menus, tables of contents, and indexes.

noscript
The HTML <noscript> element defines a section of HTML to be inserted if a script type on the page is unsupported or if scripting is currently turned off in the browser.

object
The HTML <object> element represents an external resource, which can be treated as an image, a nested browsing context, or a resource to be handled by a plugin.

ol
The HTML <ol> element represents an ordered list of items, typically rendered as a numbered list.

optgroup
The HTML <optgroup> element creates a grouping of options within a <select> element.

option
The HTML <option> element is used to define an item contained in a <select>, an <optgroup>, or a <datalist> element. As such, <option> can represent menu items in popups and other lists of items in an HTML document.

output
The HTML Output element (<output>) is a container element into which a site or app can inject the results of a calculation or the outcome of a user action.

p
The HTML <p> element represents a paragraph.

param
The [HTML <param> element][param] defines parameters for an <object> element.

picture
The HTML <picture> element contains zero or more <source> elements and one <img> element to provide versions of an image for different display/device scenarios.

pre
The HTML <pre> element represents preformatted text which is to be presented exactly as written in the HTML file.

progress
The [HTML <progress> element][progress] displays an indicator showing the completion progress of a task, typically displayed as a progress bar.

q
The HTML <q> element indicates that the enclosed text is a short inline quotation. Most modern browsers implement this by surrounding the text in quotation marks.

rb
The HTML Ruby Base (<rb>) element is used to delimit the base text component of a <ruby> annotation, i.e. the text that is being annotated.

rp
The HTML Ruby Fallback Parenthesis (<rp>) element is used to provide fall-back parentheses for browsers that do not support display of ruby annotations using the <ruby> element.

rt
The HTML Ruby Text (<rt>) element specifies the ruby text component of a ruby annotation, which is used to provide pronunciation, translation, or transliteration information for East Asian typography. The <rt> element must always be contained within a <ruby> element.

rtc
The HTML Ruby Text Container (<rtc>) element embraces semantic annotations of characters presented in a ruby of <rb> elements used inside of <ruby> element. <rb> elements can have both pronunciation (<rt>) and semantic (<rtc>) annotations.

ruby
The HTML <ruby> element represents a ruby annotation. Ruby annotations are for showing pronunciation of East Asian characters.

s
The HTML <s> element renders text with a strikethrough, or a line through it. Use the <s> element to represent things that are no longer relevant or no longer accurate. However, <s> is not appropriate when indicating document edits; for that, use the <del> and <ins> elements, as appropriate.

samp
The HTML Sample Element (<samp>) is used to enclose inline text which represents sample (or quoted) output from a computer program.

script
The HTML <script> element is used to embed or reference executable code; this is typically used to embed or refer to JavaScript code.

section
The HTML <section> element represents a standalone section — which doesn't have a more specific semantic element to represent it — contained within an HTML document.

select
The HTML <select> element represents a control that provides a menu of options.

slot
The HTML <slot> element—part of the Web Components technology suite—is a placeholder inside a web component that you can fill with your own markup, which lets you create separate DOM trees and present them together.

small
The HTML <small> element represents side-comments and small print, like copyright and legal text, independent of its styled presentation. By default, it renders text within it one font-size small, such as from small to x-small.

source
The [HTML <source> element][source] specifies multiple media resources for the <picture>, the <audio> element, or the <video> element.

span
The HTML <span> element is a generic inline container for phrasing content, which does not inherently represent anything. It can be used to group elements for styling purposes (using the class or id attributes), or because they share attribute values, such as lang.

strong
The HTML Strong Importance Element (<strong>) indicates that its contents have strong importance, seriousness, or urgency. Browsers typically render the contents in bold type.

style
The HTML <style> element contains style information for a document, or part of a document.

sub
The HTML Subscript element (<sub>) specifies inline text which should be displayed as subscript for solely typographical reasons.

summary
The HTML Disclosure Summary element (<summary>) element specifies a summary, caption, or legend for a <details> element's disclosure box.

sup
The HTML Superscript element (<sup>) specifies inline text which is to be displayed as superscript for solely typographical reasons.

table
The HTML <table> element represents tabular data — that is, information presented in a two-dimensional table comprised of rows and columns of cells containing data.

tbody
The HTML Table Body element (<tbody>) encapsulates a set of table rows (<tr> elements), indicating that they comprise the body of the table (<table>).

td
The HTML <td> element defines a cell of a table that contains data. It participates in the table model.

template
The HTML Content Template (<template>) element is a mechanism for holding HTML that is not to be rendered immediately when a page is loaded but may be instantiated subsequently during runtime using JavaScript.

text
Create and mount a DOM text node. This is normally called by the [moxie::mox!] macro.

textarea
The HTML <textarea> element represents a multi-line plain-text editing control, useful when you want to allow users to enter a sizeable amount of free-form text, for example a comment on a review or feedback form.

tfoot
The HTML <tfoot> element defines a set of rows summarizing the columns of the table.

th
The HTML <th> element defines a cell as header of a group of table cells. The exact nature of this group is defined by the scope and headers attributes.

thead
The HTML <thead> element defines a set of rows defining the head of the columns of the table.

time
The HTML <time> element represents a specific period in time.

title
The HTML Title element (<title>) defines the document's title that is shown in a browser's title bar or a page's tab.

tr
The HTML <tr> element defines a row of cells in a table. The row's cells can then be established using a mix of <td> (data cell) and <th> (header cell) elements.

track
The HTML <track> element is used as a child of the media elements <audio> and <video>. It lets you specify timed text tracks (or time-based data), for example to automatically handle subtitles. The tracks are formatted in WebVTT format (.vtt files) — Web Video Text Tracks or Timed Text Markup Language (TTML).

u
The HTML Unarticulated Annotation Element (<u>) represents a span of inline text which should be rendered in a way that indicates that it has a non-textual annotation.

ul
The HTML <ul> element represents an unordered list of items, typically rendered as a bulleted list.

var
The HTML Variable element (<var>) represents the name of a variable in a mathematical expression or a programming context.

video
The HTML Video element (<video>) embeds a media player which supports video playback into the document.

wbr
The HTML <wbr> element represents a word break opportunity—a position within text where the browser may optionally break a line, though its line-breaking rules would not otherwise create a break at that location.

Structs

MemoElement
A topologically-nested "incremental smart pointer" for an HTML element.

Enums

Node
A Node in the augmented DOM.

Traits

Dom
A value which implements a subset of the web's document object model.

JsCast
A trait for checked and unchecked casting between JS types.

Functions

boot
The "boot sequence" for a moxie-dom instance creates a [crate::embed::WebRuntime] with the provided arguments and begins scheduling its execution with requestAnimationFrame on state changes.

document
Returns the current document. Panics if called outside a web document context.

render_html
Runs the provided closure once and produces a prettified HTML string from the contents.

window
Returns the current window. Panics if no window is available.