Module moxie_dom::event

Event types.

Structs

Abort
The loading of a resource has been aborted. MDN documentation

AbortProgress
Progression has been terminated (not due to an error). MDN documentation

AbortTransaction
A transaction has been aborted. MDN documentation

AfterPrint
The associated document has started printing or the print preview has been closed. MDN documentation

AnimationCancel
A CSS animation has aborted. MDN documentation

AnimationEnd
A CSS animation has completed. MDN documentation

AnimationIteration
A CSS animation is ticked. MDN documentation

AnimationRepeat
A SMIL animation element is repeated. MDN documentation

AnimationStart
A CSS animation has started. MDN documentation

AppInstalled
A web application is successfully installed as a progressive web app. MDN documentation

AudioComplete
The rendering of an OfflineAudioContext is terminated. MDN documentation

AudioEnd
The user agent has finished capturing audio for speech recognition. MDN documentation

AudioEnded
Playback has stopped because the end of the media was reached. MDN documentation

AudioProcess
The input buffer of a ScriptProcessorNode is ready to be processed. MDN documentation

AudioStart
The user agent has started to capture audio for speech recognition. MDN documentation

BeforePrint
The associated document is about to be printed or previewed for printing. MDN documentation

BeforeUnload
The window, the document and its resources are about to be unloaded. MDN documentation

Blur
An element has lost focus (does not bubble). MDN documentation

CanPlay
The user agent can play the media, but estimates that not enough data has been loaded to play the media up to its end without having to stop for further buffering of content. MDN documentation

CanPlayThrough
The user agent can play the media up to its end without having to stop for further buffering of content. MDN documentation

Change
The change event is fired for <textarea> elements when a change to the element's value is committed by the user. MDN documentation

ChargingChange
The battery begins or stops charging. MDN documentation

ChargingTime
The chargingTime attribute has been updated. MDN documentation

ChildMessage
A message is received from a child (i)frame or a parent window. MDN documentation

Click
A pointing device button has been pressed and released on an element. MDN documentation

CloseWebsocket
A WebSocket connection has been closed. MDN documentation

CompositionEnd
The composition of a passage of text has been completed or canceled. MDN documentation

CompositionStart
The composition of a passage of text is prepared (similar to keydown for a keyboard input, but works with other inputs such as speech recognition). MDN documentation

CompositionUpdate
A character is added to a passage of text being composed. MDN documentation

ConnectionBlocked
An open connection to a database is blocking a versionchange transaction on the same database. MDN documentation

ContextMenu
The right button of the mouse is clicked (before the context menu is displayed). MDN documentation

ContextMenuShow
A contextmenu event was fired on/bubbled to an element that has a contextmenu attribute. MDN documentation

Cp
The text selection has been added to the clipboard. MDN documentation

Cut
The text selection has been removed from the document and added to the clipboard. MDN documentation

DataLoaded
The first frame of the media has finished loading. MDN documentation

DeviceChange
A media device such as a camera, microphone, or speaker is connected or removed from the system. MDN documentation

DeviceMotion
Fresh data is available from a motion sensor. MDN documentation

DeviceOrientation
Fresh data is available from an orientation sensor. MDN documentation

DischargingTime
The dischargingTime attribute has been updated. MDN documentation

DomContentLoaded
The document has finished loading (but not its dependent resources). MDN documentation

DoubleClick
A pointing device button is clicked twice on an element. MDN documentation

Drag
An element or text selection is being dragged (every 350ms). MDN documentation

DragEnd
A drag operation is being ended (by releasing a mouse button or hitting the escape key). MDN documentation

DragEnter
A dragged element or text selection enters a valid drop target. MDN documentation

DragLeave
A dragged element or text selection leaves a valid drop target. MDN documentation

DragOver
An element or text selection is being dragged over a valid drop target (every 350ms). MDN documentation

DragStart
The user starts dragging an element or text selection. MDN documentation

Dropped
An element is dropped on a valid drop target. MDN documentation

DurationChange
The duration attribute has been updated. MDN documentation

Emptied
The media has become empty; for example, this event is sent if the media has already been loaded (or partially loaded), and the load() method is called to reload it. MDN documentation

EventSourceError
An event source connection has been failed. MDN documentation

EventSourceMessage
A message is received through an event source. MDN documentation

EventSourceOpen
An event source connection has been established. MDN documentation

Focus
An element has received focus (does not bubble). MDN documentation

FocusIn
An element is about to receive focus (bubbles). MDN documentation

FocusOut
An element is about to lose focus (bubbles). MDN documentation

FormReset
A form is reset. MDN documentation

FullscreenChange
An element was turned to fullscreen mode or back to normal mode. MDN documentation

FullscreenError
It was impossible to switch to fullscreen mode for technical reasons or because the permission was denied. MDN documentation

GamepadConnected
A gamepad has been connected. MDN documentation

GamepadDisconnected
A gamepad has been disconnected. MDN documentation

GotPointerCapture
Element receives pointer capture. MDN documentation

HashChange
The fragment identifier of the URL has changed (the part of the URL after the #). MDN documentation

Input
The value of an element changes or the content of an element with the attribute contenteditable is modified.

Invalid
A submittable element has been checked and doesn't satisfy its constraints. MDN documentation

KeyDown
A key is pressed down. MDN documentation

KeyUp
A key is released. MDN documentation

LanguageChange
The user's preferred languages have changed. MDN documentation

LevelChange
The level attribute has been updated. MDN documentation

LoadEnd
Progress has stopped (after "error", "abort" or "load" have been dispatched). MDN documentation

LoadStart
Progress has begun. MDN documentation

LostPointerCapture
Element lost pointer capture. MDN documentation

MessageError
A message error is raised when a message is received by an object. MDN documentation

MetadataLoaded
The metadata has been loaded. MDN documentation

MouseDown
A pointing device button (usually a mouse) is pressed on an element. MDN documentation

MouseEnter
A pointing device is moved onto the element that has the listener attached. MDN documentation

MouseLeave
A pointing device is moved off the element that has the listener attached. MDN documentation

MouseMove
A pointing device is moved over an element. MDN documentation

MouseOut
A pointing device is moved off the element that has the listener attached or off one of its children. MDN documentation

MouseOver
A pointing device is moved onto the element that has the listener attached or onto one of its children. MDN documentation

MouseUp
A pointing device button is released over an element. MDN documentation

NotificationClick
A system notification spawned by ServiceWorkerRegistration.showNotification() has been clicked. MDN documentation

Offline
The browser has lost access to the network. MDN documentation

Online
The browser has gained access to the network (but particular websites might be unreachable). MDN documentation

OrientationChange
The orientation of the device (portrait/landscape) has changed. MDN documentation

PageHide
A session history entry is being traversed from. MDN documentation

PageShow
A session history entry is being traversed to. MDN documentation

Paste
Data has been transferred from the system clipboard to the document. MDN documentation

Pause
Playback has been paused. MDN documentation

Play
Playback has begun. MDN documentation

PlaybackEnded
Playback has stopped because the end of the media was reached. MDN documentation

PlaybackRateChange
The playback rate has changed. MDN documentation

Playing
Playback is ready to start after having been paused or delayed due to lack of data. MDN documentation

PointerCancel
The pointer is unlikely to produce any more events. MDN documentation

PointerDown
The pointer enters the active buttons state. MDN documentation

PointerEnter
Pointing device is moved inside the hit-testing boundary. MDN documentation

PointerLeave
Pointing device is moved out of the hit-testing boundary. MDN documentation

PointerLockChange
The pointer was locked or released. MDN documentation

PointerLockError
It was impossible to lock the pointer for technical reasons or because the permission was denied. MDN documentation

PointerMove
The pointer changed coordinates. MDN documentation

PointerOut
The pointing device moved out of hit-testing boundary or leaves detectable hover range. MDN documentation

PointerOver
The pointing device is moved into the hit-testing boundary. MDN documentation

PointerUp
The pointer leaves the active buttons state. MDN documentation

PopState
A session history entry is being navigated to (in certain cases). MDN documentation

Progress
In progress. MDN documentation

ProgressError
Progression has failed. MDN documentation

ProgressLoad
Progression has been successful. MDN documentation

Push
A Service Worker has received a push message. MDN documentation

PushSubscriptionChange
A PushSubscription has expired. MDN documentation

ReadyStateChange
The readyState attribute of a document has changed. MDN documentation

RequestError
A request caused an error and failed. MDN documentation

ResourceError
A resource failed to load. MDN documentation

ResourceLoad
A resource and its dependent resources have finished loading. MDN documentation

Scroll
The document view or an element has been scrolled. MDN documentation

Seeked
A seek operation completed. MDN documentation

Seeking
A seek operation began. MDN documentation

Select
Some text is being selected. MDN documentation

SelectionChange
The selection in the document has been changed. MDN documentation

SelectionStart
A selection just started. MDN documentation

ServiceWorkerMessage
A message is received from a service worker, or a message is received in a service worker from another context. MDN documentation

SlotChange
The node contents of a <slot> have changed. MDN documentation

SoundEnd
Any sound — recognisable speech or not — has stopped being detected. MDN documentation

SoundStart
Any sound — recognisable speech or not — has been detected. MDN documentation

SpeechBoundary
The spoken utterance reaches a word or sentence boundary. MDN documentation

SpeechEnd
Speech recognised by the speech recognition service has stopped being detected. MDN documentation

SpeechError
An error occurs that prevents the utterance from being successfully spoken. MDN documentation

SpeechMark
The spoken utterance reaches a named SSML "mark" tag. MDN documentation

SpeechPause
The utterance is paused part way through. MDN documentation

SpeechRecognitionEnd
The speech recognition service has disconnected. MDN documentation

SpeechRecognitionError
A speech recognition error occurs. MDN documentation

SpeechRecognitionNoMatch
The speech recognition service returns a final result with no significant recognition. MDN documentation

SpeechRecognitionResult
The speech recognition service returns a result — a word or phrase has been positively recognized and this has been communicated back to the app. MDN documentation

SpeechRecognitionStart
The speech recognition service has begun listening to incoming audio with intent to recognize grammars associated with the current SpeechRecognition. MDN documentation

SpeechRecognized
Sound that is recognised by the speech recognition service as speech has been detected. [MDN documentation][mdn]

SpeechResume
A paused utterance is resumed. MDN documentation

SpeechStart
The utterance has begun to be spoken. MDN documentation

SpeechSynthesisEnd
The utterance has finished being spoken. MDN documentation

Stalled
The user agent is trying to fetch media data, but data is unexpectedly not forthcoming. MDN documentation

Storage
A storage area (localStorage or sessionStorage) has changed. MDN documentation

Submit
A form is submitted. MDN documentation

Success
A request successfully completed. MDN documentation

Suspend
Media data loading has been suspended. MDN documentation

SvgAbort
Page loading has been stopped before the SVG was loaded. MDN documentation

SvgAnimationBegin
A SMIL animation element begins. MDN documentation

SvgAnimationEnd
A SMIL animation element ends. MDN documentation

SvgError
An error has occurred before the SVG was loaded. MDN documentation

SvgLoad
An SVG document has been loaded and parsed. MDN documentation

SvgResize
An SVG document is being resized. MDN documentation

SvgScroll
An SVG document is being scrolled. MDN documentation

SvgUnload
An SVG document has been removed from a window or frame. MDN documentation

SvgZoom
An SVG document is being zoomed. MDN documentation

TimeUpdate
The time indicated by the currentTime attribute has been updated. MDN documentation

Timeout
The timeout event is fired when progression is terminated due to preset time expiring. MDN documentation

TimingBufferFull
The browser's resource timing buffer is full. MDN documentation

TouchCancel
A touch point has been disrupted in an implementation-specific manners (too many touch points for example). MDN documentation

TouchEnd
A touch point is removed from the touch surface. MDN documentation

TouchMove
A touch point is moved along the touch surface. MDN documentation

TouchStart
A touch point is placed on the touch surface. MDN documentation

TransactionComplete
A transaction successfully completed. MDN documentation

TransactionVersionChange
A versionchange transaction completed. MDN documentation

TransitionEnd
A CSS transition has completed. MDN documentation

Unload
The document or a dependent resource is being unloaded. MDN documentation

UpgradeNeeded
An attempt was made to open a database with a version number higher than its current version. A versionchange transaction has been created. MDN documentation

UserProximity
Fresh data is available from a proximity sensor (indicates whether the nearby object is near the device or not). MDN documentation

ViewResize
The document view has been resized. MDN documentation

VisibilityChange
The content of a tab has become visible or has been hidden. MDN documentation

VoicesChanged
Fires when the list of SpeechSynthesisVoice objects that would be returned by the SpeechSynthesis.getVoices() method has changed. MDN documentation

VolumeChange
The volume has changed. MDN documentation

Waiting
Playback has stopped because of a temporary lack of data. MDN documentation

WebsocketError
A WebSocket connection has been closed with prejudice (some data couldn't be sent for example). MDN documentation

WebsocketMessage
A message is received through a WebSocket. MDN documentation

WebsocketOpen
A WebSocket connection has been established. MDN documentation

Wheel
A wheel button of a pointing device is rotated in any direction. MDN documentation

WorkerMessage
A message is received from a Web Worker. MDN documentation

Traits

Event
An event that can be received as the first argument to a handler callback.